# openapi.yml
openapi: 3.0.3
info:
  title: S3Desk API
  version: 0.1.0
  description: |
    Draft API spec for S3Desk, a local-only (127.0.0.1) dashboard targeting S3-compatible object storage.
    - Metadata/browsing and downloads are powered by `rclone`.
    - Bulk transfers run as Jobs powered by `rclone`.
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Default local server
...
paths:
...
  /server/backup:
    get:
      tags: [System]
      summary: Download a server backup bundle
      description: |
        Exports a `.tar.gz` server backup bundle containing the sqlite snapshot and a
        selected slice of `DATA_DIR` runtime state. Use `scope=full` for a full local
        backup, or `scope=cache_metadata` for a lighter bundle containing metadata and
        cache state such as thumbnails. Use `scope=portable` to export a logical
        database-neutral migration bundle for portable import flows. Use
        `confidentiality=encrypted` to encrypt the payload with
        a non-empty `X-S3Desk-Backup-Password` when supplied, otherwise with the
        current `ENCRYPTION_KEY`, while keeping the outer bundle metadata clear.
        Restore/import endpoints do not read the export header; supply password-protected
        bundles through their multipart `password` field or matching UI input. During
        restore/import, a non-empty multipart password is used before destination
        server `ENCRYPTION_KEY` fallback. Snapshot-style export currently supports
        sqlite-backed servers only, while portable export supports sqlite and postgres
        source deployments.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
        - name: X-S3Desk-Backup-Password
          in: header
          required: false
          description: Optional operator-supplied export password used when `confidentiality=encrypted`. Restore/import endpoints do not read this header; provide the same value in their multipart `password` field when processing password-protected bundles.
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
            enum: [full, cache_metadata, portable]
            default: full
          description: Selects whether the bundle contains a sqlite snapshot export (`full`, `cache_metadata`) or a logical portable migration bundle (`portable`).
        - name: includeThumbnails
          in: query
          schema:
            type: boolean
            default: true
          description: When `scope=portable`, includes thumbnail cache assets in the bundle.
        - name: confidentiality
          in: query
          schema:
            type: string
            enum: [clear, encrypted]
            default: clear
          description: Encrypts the bundle payload with a non-empty `X-S3Desk-Backup-Password` when supplied, otherwise with the current `ENCRYPTION_KEY`, while leaving the outer tar.gz manifest readable.
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              description: Suggested filename for the backup bundle.
              schema: { type: string }
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
  /server/backup/transfer:
    post:
      tags: [System]
      summary: Export a server backup to remote storage
      description: Creates the same validated backup bundle as `GET /server/backup`, then copies it to an existing object-storage profile, an FTP server, or an NFS path already mounted on the server. NFS paths are constrained by `ALLOWED_LOCAL_DIRS`. FTP is plain FTP; use encrypted backup confidentiality when the network is not trusted.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerBackupTransferRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerBackupTransferResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /server/restore/transfer:
    post:
      tags: [System]
      summary: Fetch and stage a server backup from remote storage
      description: Fetches one backup bundle from an existing object-storage profile, FTP server, or mounted NFS path, enforces the restore size limit, and passes it through the same checksum, signature, encryption, preflight, and stage-only restore flow as `POST /server/restore`.
      parameters:
        - $ref: "#/components/parameters/XApiToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerRestoreTransferRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRestoreResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "413":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  parameters:
...
    XApiToken:
      name: X-Api-Token
      in: header
      required: false
      description: Optional local API token to mitigate localhost/CSRF style attacks.
      schema: { type: string }
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            normalizedError:
              $ref: "#/components/schemas/NormalizedError"
            details: { type: object, additionalProperties: true }
    NormalizedError:
      type: object
      description: Provider-agnostic error classifier (primarily derived from rclone stderr).
      required: [code, retryable]
      properties:
        code:
          $ref: "#/components/schemas/NormalizedErrorCode"
        retryable:
          type: boolean
    NormalizedErrorCode:
      type: string
      enum:
        - invalid_credentials
        - access_denied
        - not_found
        - rate_limited
        - network_error
        - invalid_config
        - signature_mismatch
        - request_time_skewed
        - conflict
        - upstream_timeout
        - endpoint_unreachable
        - canceled
        - unknown
...
    ServerMigrationManifest:
      type: object
      required: [format, bundleKind, createdAt, appVersion, dbBackend, encryptionEnabled]
      properties:
        format: { type: string }
        bundleKind:
          type: string
          enum: [full, cache_metadata, portable]
        formatVersion: { type: integer }
        confidentialityMode:
          type: string
          enum: [clear, encrypted]
        createdAt: { type: string, format: date-time }
        appVersion: { type: string }
        dbBackend: { type: string }
        schemaVersion: { type: integer }
        encryptionEnabled: { type: boolean }
        encryptionKeyHint: { type: string }
        entries:
          type: array
          items: { type: string }
        entities:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ServerMigrationEntityManifest"
        assets:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ServerMigrationAssetManifest"
        payloadFileCount: { type: integer }
        payloadBytes: { type: integer, format: int64 }
        payloadSha256: { type: string }
        warnings:
          type: array
          items: { type: string }
    ServerMigrationEntityManifest:
      type: object
      required: [count, sha256]
      properties:
        count: { type: integer }
        sha256: { type: string }
    ServerMigrationAssetManifest:
      type: object
      properties:
        fileCount: { type: integer }
        bytes: { type: integer, format: int64 }
        sha256: { type: string }
    ServerRestoreResponse:
      type: object
      required: [manifest, validation, stagingDir, restartRequired, nextSteps, applyPlan]
      properties:
        manifest:
          $ref: "#/components/schemas/ServerMigrationManifest"
        validation:
          $ref: "#/components/schemas/ServerRestoreValidation"
        stagingDir: { type: string }
        restartRequired: { type: boolean }
        nextSteps:
          type: array
          items: { type: string }
        applyPlan:
          type: array
          items: { type: string }
        helperCommand: { type: string }
        warnings:
          type: array
          items: { type: string }
    ServerBackupTransferLocation:
      type: object
      required: [protocol, path]
      properties:
        protocol:
          type: string
          enum: [object_storage, ftp, nfs]
        profileId:
          type: string
          description: Existing object-storage profile ID; required for `object_storage`.
        bucket:
          type: string
          description: Object-storage bucket or container; required for `object_storage`.
        path:
          type: string
          description: Object key, FTP file path, or mounted NFS file path. On export, a trailing slash appends the generated backup filename.
        host:
          type: string
          description: FTP host; required for `ftp`.
        port:
          type: integer
          minimum: 1
          maximum: 65535
          default: 21
        username:
          type: string
          description: FTP username; required for `ftp`.
        password:
          type: string
          description: FTP password. It is passed to rclone through stdin-protected temporary configuration and is never returned.
    ServerBackupTransferRequest:
      type: object
      required: [location]
      properties:
        scope:
          type: string
          enum: [full, cache_metadata, portable]
          default: full
        confidentiality:
          type: string
          enum: [clear, encrypted]
          default: clear
        backupPassword:
          type: string
          description: Optional operator password for an encrypted bundle.
        includeThumbnails:
          type: boolean
        location:
          $ref: "#/components/schemas/ServerBackupTransferLocation"
    ServerRestoreTransferRequest:
      type: object
      required: [location]
      properties:
        backupPassword:
          type: string
          description: Optional password used to decrypt the fetched bundle.
        location:
          $ref: "#/components/schemas/ServerBackupTransferLocation"
    ServerBackupTransferResponse:
      type: object
      required: [protocol, location, filename, sizeBytes]
      properties:
        protocol:
          type: string
          enum: [object_storage, ftp, nfs]
        location: { type: string }
        filename: { type: string }
        sizeBytes: { type: integer, format: int64 }
    ServerRestoreValidation:
      type: object
      required: [preflightChecked, payloadChecksumPresent, payloadChecksumVerified, payloadSignaturePresent, payloadSignatureVerified]
      properties:
        preflightChecked: { type: boolean }
        diskFreeBytesBefore: { type: integer, format: int64 }
        payloadFileCount: { type: integer }
        payloadBytes: { type: integer, format: int64 }
        payloadChecksumPresent: { type: boolean }
        payloadChecksumVerified: { type: boolean }
        payloadSignaturePresent: { type: boolean }
        payloadSignatureVerified: { type: boolean }
        payloadEncryptionPresent: { type: boolean }
        payloadEncryptionDecrypted: { type: boolean }
...